refactor: move prModal to class
